boomai/data/requirements.txt
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.0
PyGithub==2.5.0
